+++ deltafi-ui/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Changed
+
+- Added delete button to System Snapshots page (#295)
+
 ## [0.99.3]
 
 ### Changed